move definitions to xml-io.lua
