+++ xmlua.rockspec
@@ -64,6 +64,7 @@ build = {
     ["xmlua.libxml2.parser"] = "xmlua/libxml2/parser.lua",
     ["xmlua.libxml2.tree"] = "xmlua/libxml2/tree.lua",
     ["xmlua.libxml2.valid"] = "xmlua/libxml2/valid.lua",
+    ["xmlua.libxml2.xml-io"] = "xmlua/libxml2/xml-io.lua",
     ["xmlua.libxml2.xmlerror"] = "xmlua/libxml2/xmlerror.lua",
     ["xmlua.libxml2.xmlsave"] = "xmlua/libxml2/xmlsave.lua",
     ["xmlua.libxml2.xmlstring"] = "xmlua/libxml2/xmlstring.lua",
